extract execution stages from dag to schedule
Dags that have nodes with multiple dependents need to be executed in stages. Dependent nodes are scheduled only after the dependencies are executed. This is done to avoid possible duplicate execution of dependencies.

This also lets us determine cases where it will be cheaper to execute a dependency multiple times (based on the cost of execution of dependency vs the cost of communicating the result to multiple nodes), and duplicate certain tasks. (TBD later)